Main view fix

+++ app/src/main/java/pl/kostka/restaurantclient/model/HomeElement.kt
@@ -2,7 +2,7 @@ package pl.kostka.restaurantclient.model
 
 import android.graphics.drawable.Icon
 
-class HomeElement(val id: Long? = null,
+class HomeElement(val id: Int? = null,
               val name: String? = null,
               val description: String? = null,
               val link: String? = null,
